# Improve read_multi_entries performance for numerous keys